requirements.txt
autopep8==2.0.2
cssmin==0.2.0
Flask==2.3.1
Flask-Assets==2.0
jsmin==3.0.1
opencv-python==4.7.0.72
pyscss==1.4.0
python-dotenv==1.0.0
